# internal/api/openapi.yaml
openapi: 3.1.0
info:
  title: MoxApp API
  description: |
    High-performance concurrent HTTP load testing tool with DNS timing metrics.
    
    This API provides endpoints for:
    - **Outgoing Traffic Management**: Configure and control external API calls
    - **Incoming Traffic Simulation**: Define mock routes with configurable response patterns
    - **Metrics & Monitoring**: Real-time metrics collection and health monitoring
    - **Scheduler Control**: Pause, resume, or emergency stop the load test
  version: 1.0.0
  contact:
    name: MoxApp Team
  license:
    name: Proprietary
    identifier: LicenseRef-MoxApp
servers:
  - url: __SERVER_URL__
    description: Current server
...
paths:
...
  /api/incoming/routes:
    get:
      tags:
        - Incoming Routes
      summary: List all incoming routes
      description: Returns all configured incoming route simulations
      operationId: listIncomingRoutes
      responses:
        '200':
          description: List of incoming routes
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IncomingRouteListResponse'
        '503':
          description: Incoming routes not configured
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    post:
      tags:
        - Incoming Routes
      summary: Create incoming route
      description: Creates a new incoming route simulation
      operationId: createIncomingRoute
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/IncomingRouteRequest'
      responses:
        '201':
          description: Incoming route created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IncomingRouteCreateResponse'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '503':
          description: Incoming routes manager not available
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/incoming/routes/{name}:
    get:
      tags:
        - Incoming Routes
      summary: Get incoming route by name
      description: Returns a specific incoming route configuration
      operationId: getIncomingRoute
      parameters:
        - name: name
          in: path
          required: true
          description: Route name
          schema:
            type: string
      responses:
        '200':
          description: Incoming route configuration
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IncomingRoute'
        '400':
          description: Route name required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Route not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '503':
          description: Incoming routes not configured
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    put:
      tags:
        - Incoming Routes
      summary: Update incoming route
      description: Updates an existing incoming route configuration
      operationId: updateIncomingRoute
      parameters:
        - name: name
          in: path
          required: true
          description: Route name
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/IncomingRouteRequest'
      responses:
        '200':
          description: Route updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IncomingRouteUpdateResponse'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Route not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '503':
          description: Incoming routes manager not available
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      tags:
        - Incoming Routes
      summary: Delete incoming route
      description: Deletes an incoming route by name
      operationId: deleteIncomingRoute
      parameters:
        - name: name
          in: path
          required: true
          description: Route name
          schema:
            type: string
      responses:
        '200':
          description: Route deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IncomingRouteDeleteResponse'
        '400':
          description: Route name required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Route not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '503':
          description: Incoming routes manager not available
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/incoming/routes/reload:
    post:
      tags:
        - Incoming Routes
      summary: Reload routes from config
      description: |
        Reloads all incoming routes from the static configuration file.
        **Warning**: This replaces all runtime-created routes with the static configuration.
      operationId: reloadIncomingRoutes
      responses:
        '200':
          description: Routes reloaded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IncomingRouteReloadResponse'
        '400':
          description: No config file path configured
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Failed to reload config
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '503':
          description: Incoming routes manager not available
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /api/incoming/control/route:
    post:
      tags:
        - Incoming Control
      summary: Enable/disable specific route
      description: Enables or disables a specific incoming route
      operationId: setIncomingRouteEnabled
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
                - enabled
              properties:
                name:
                  type: string
                  description: Route name
                enabled:
                  type: boolean
                  description: Enable or disable the route
      responses:
        '200':
          description: Route status updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IncomingRouteControlResponse'
        '400':
          description: Invalid request or route name required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Route not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '503':
          description: Incoming routes manager not available
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  schemas:
...
    IncomingRoute:
      type: object
      properties:
        name:
          type: string
          description: Unique route identifier
          example: call_info
        path:
          type: string
          description: URL path (supports prefix matching)
          example: /api/call_info
        method:
          type: string
          description: HTTP method or "*" for any
          example: GET
        enabled:
          type: boolean
          description: Whether route is active
          example: true
        responses:
          type: array
          items:
            $ref: '#/components/schemas/IncomingResponseConfig'
    IncomingResponseConfig:
      type: object
      required:
        - status
        - share
        - min_response_ms
        - max_response_ms
      properties:
        status:
          type: integer
          description: HTTP status code (100-599)
          example: 200
        share:
          type: number
          format: float
          description: Probability weight (0.0-1.0, all must sum to 1.0)
          example: 0.9
        min_response_ms:
          type: integer
          description: Minimum simulated response time in milliseconds
          example: 100
        max_response_ms:
          type: integer
          description: Maximum simulated response time in milliseconds
          example: 300
    IncomingRouteRequest:
      type: object
      required:
        - name
        - path
        - method
        - responses
      properties:
        name:
          type: string
          description: Unique route identifier
          example: user_service
        path:
          type: string
          description: URL path (must start with /)
          example: /api/users
        method:
          type: string
          description: HTTP method (GET, POST, PUT, DELETE, PATCH, HEAD, OPTIONS) or "*" for any
          example: GET
        enabled:
          type: boolean
          description: Whether route is active
          default: true
        responses:
          type: array
          minItems: 1
          items:
            $ref: '#/components/schemas/IncomingResponseConfig'
          description: Response configurations (shares must sum to 1.0)
    IncomingRouteListResponse:
      type: object
      properties:
        enabled:
          type: boolean
        count:
          type: integer
        routes:
          type: array
          items:
            $ref: '#/components/schemas/IncomingRoute'
        sim_prefix:
          type: string
          example: /sim
    IncomingRouteCreateResponse:
      type: object
      properties:
        message:
          type: string
          example: incoming route created
        route:
          $ref: '#/components/schemas/IncomingRoute'
        sim_path:
          type: string
          example: /sim/api/users
    IncomingRouteUpdateResponse:
      type: object
      properties:
        message:
          type: string
          example: incoming route updated
        route:
          $ref: '#/components/schemas/IncomingRoute'
        sim_path:
          type: string
          example: /sim/api/users
    IncomingRouteDeleteResponse:
      type: object
      properties:
        message:
          type: string
          example: incoming route deleted
        name:
          type: string
    IncomingRouteReloadResponse:
      type: object
      properties:
        message:
          type: string
          example: incoming routes reloaded from config
        config_path:
          type: string
        count:
          type: integer
        routes:
          type: array
          items:
            $ref: '#/components/schemas/IncomingRoute'
...
    IncomingRouteControlResponse:
      type: object
      properties:
        message:
          type: string
          example: incoming route status updated
        name:
          type: string
        enabled:
          type: boolean
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          description: Error message
...